--- app/templates/posts/index.hbs
@@ -1,40 +1,0 @@
-<section class="content-list">
-    <ol class="posts-list">
-        {{#each model as |post|}}
-            {{gh-posts-list-item
-                post=post
-                active=(eq post selectedPost)
-                onDoubleClick="openEditor"
-                data-test-posts-list-item-id=post.id}}
-        {{else}}
-            <li class="no-posts-box">
-                <div class="no-posts">
-                    {{#if showingAll}}
-                        <h3>You Haven't Written Any Posts Yet!</h3>
-                        {{#link-to "editor.new"}}<button type="button" class="gh-btn gh-btn-green gh-btn-lg"><span>Write a new Post</span></button>{{/link-to}}
-                    {{else}}
-                        <h3>No posts that match the current filter</h3>
-                        {{#link-to "posts.index" (query-params type=null)}}<button type="button" class="gh-btn gh-btn-lg"><span>Show all posts</span></button>{{/link-to}}
-                    {{/if}}
-                </div>
-            </li>
-        {{/each}}
-    </ol>
-
-    {{infinity-loader
-        infinityModel=model
-        scrollable=".gh-main"
-        triggerOffset=1000}}
-</section>
-
-{{#if showDeletePostModal}}
-    {{gh-fullscreen-modal "delete-post"
-                          model=(hash
-                              post=selectedPost
-                              onSuccess=(route-action 'onPostDeletion')
-                          )
-                          close=(action "toggleDeletePostModal")
-                          modifier="action wide"}}
-{{/if}}
-
-{{outlet}}
